Rename modules to follow PEP 8, fix up doc string
