+++ worlds/spyro/locations.py
