+++ worlds/spyro/rules.py
@@ -1,9 +1,5 @@
 from typing_extensions import TYPE_CHECKING
 
-from BaseClasses import Entrance
-from worlds.generic.Rules import add_rule
-from .Regions import hub_names
-
 if TYPE_CHECKING:
     from . import SpyroWorld
